netdev-offload: Add xdp flow api provider
This provider offloads classifier to software XDP (not hardware at this
point).

It works only when a custom XDP object is loaded by afxdp netdev.

When the custom object meets certain requirements, the provider
determines that it supports classifier and flows can be offloaded to the
program. The requirements are like:

- The object has an array-of-map named "flow_table".
- The object has an hash-table named "subtbl_templates".
- The object has an array named "subtbl_masks".

etc. For details please check xdp_preload() in netdev-afxdp.c and
netdev_xdp_init_flow_api() in lib/netdev-offload-xdp.c.

Relation between maps of the object is:
- Entries of array-of-map "flow_table" are hash-tables (i.e. subtables).
- Array "subtbl_masks" has the same number of entries as "flow_tables"
  and has mask information (minimask) of subtables.  Each entry also has
  one-byte "next" field to implement linked-list with this array.
- Subtables (hash-tables, entries of "flow_table") has miniflow key and
  their value is actions.

The expected behavior of the XDP program is:
1. Parse received packets and make miniflow.
2. Iterate each mask in "subtbl_masks" and mask the packet miniflow with
   subtable minimask.
3. Look up the masked key in subtable hash-table (entry of "flow_table").
4. If an entry matches the key, execute its actions.

The provider uses BTF information to determine which key/action is
supported by the program. Please see probe_supported_keys() and
probe_supported_actions() in netdev-offload-xdp.c.

In the future it may be possible to offload classifier to SmartNIC
through XDP, but it will require map-in-map support in NIC HW or
TCAM-like map support in BPF and HW which currently does not exist.

Signed-off-by: Toshiaki Makita <toshiaki.makita1@gmail.com>