tun_metadata: Use OVS_ALIGNED_VAR to align opts field
In the following commit I want to include flow.h from BPF programs, but
with "clang -target bpf" with clang 9, UINTPTR_MAX somehow equals to
UINT32_MAX while the size of pointer is 8. As a result unnecessary pad
is added in struct tun_metadata in bpf code and cannot build bpf code
due to build assertion.

Use OVS_ALIGNED_VAR instead of pad. This works for clang BPF as well as
other compilers and architectures.

Signed-off-by: Toshiaki Makita <toshiaki.makita1@gmail.com>