bpf: Fix action loop verifier errors
As there was only OUTPUT action implemented, action loop was not verified
as a loop. As a result, I was not aware of the problem that verifier
cannot accept current loop logic which does not look like a bounded loop.

Rewrote the action loop from scratch to pass the verifier.

Signed-off-by: Toshiaki Makita <toshiaki.makita1@gmail.com>